frontend-3 Fix the Twitter Panel
