+++ src/index.html
@@ -10,11 +10,11 @@
   <link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
   <!-- jQuery -->
 <!--  <script src = "https://code.jquery.com/jquery-1.10.2.js"></script>-->
-  <script src = "https://code.jquery.com/ui/1.10.4/jquery-ui.js"></script>
-  <link href = "https://code.jquery.com/ui/1.10.4/themes/ui-lightness/jquery-ui.css" rel = "stylesheet">
-  <link rel="stylesheet" href="https://unpkg.com/leaflet@1.3.1/dist/leaflet.css"
-        integrity="sha512-Rksm5RenBEKSKFjgI3a41vrjkw4EVPlJ3+OiI65vTjIdo9brlAacEuKOiQ5OFh7cOI1bkDwLqdLw3Zg0cRJAAQ=="
-        crossorigin=""/>
+<!--  <script src = "https://code.jquery.com/ui/1.10.4/jquery-ui.js"></script>-->
+<!--  <link href = "https://code.jquery.com/ui/1.10.4/themes/ui-lightness/jquery-ui.css" rel = "stylesheet">-->
+<!--  <link rel="stylesheet" href="https://unpkg.com/leaflet@1.3.1/dist/leaflet.css"-->
+<!--        integrity="sha512-Rksm5RenBEKSKFjgI3a41vrjkw4EVPlJ3+OiI65vTjIdo9brlAacEuKOiQ5OFh7cOI1bkDwLqdLw3Zg0cRJAAQ=="-->
+<!--        crossorigin=""/>-->
 
   <!-- leaflet style sheet -->
 <!--  <link rel="stylesheet" href="https://unpkg.com/leaflet@1.3.1/dist/leaflet.css"-->
@@ -25,14 +25,14 @@
 <!--          integrity="sha512-/Nsx9X4HebavoBvEBuyp3I7od5tA0UzAxs+j83KgC8PU0kgB4XiK4Lfe4y4cgBtaRJQEIFCW+oC506aPT2L1zw=="-->
 <!--          crossorigin=""></script>-->
 <!--  &lt;!&ndash; Nice Tweets &ndash;&gt;-->
-<!--  <script>-->
-<!--    window.twttr = (function (d,s,id) {-->
-<!--      var t, js, fjs = d.getElementsByTagName(s)[0];-->
-<!--      if (d.getElementById(id)) return; js=d.createElement(s); js.id=id;-->
-<!--      js.src="https://platform.twitter.com/widgets.js"; fjs.parentNode.insertBefore(js, fjs);-->
-<!--      return window.twttr || (t = { _e: [], ready: function(f){ t._e.push(f) } });-->
-<!--    }(document, "script", "twitter-wjs"));-->
-<!--  </script>-->
+  <script>
+    window.twttr = (function (d,s,id) {
+      var t, js, fjs = d.getElementsByTagName(s)[0];
+      if (d.getElementById(id)) return; js=d.createElement(s); js.id=id;
+      js.src="https://platform.twitter.com/widgets.js"; fjs.parentNode.insertBefore(js, fjs);
+      return window.twttr || (t = { _e: [], ready: function(f){ t._e.push(f) } });
+    }(document, "script", "twitter-wjs"));
+  </script>
 
 </head>
 <body>
